fixBugsForProfileAndGFHAHA
