+++ src/containers/Member/Profile/Profile.js
@@ -1,10 +1,12 @@
 import { CCard, CCardBody, CCardHeader, CCol, CRow } from "@coreui/react";
 import { AccountContext } from "hooks/context";
 import { useContext } from "react";
+import { registFormat } from "utils/date";
 
 const ModifyCard = () => {
   var account = useContext(AccountContext);
   if (!account) return null;
+
   return (
     <CCardBody>
       <CRow>
@@ -22,19 +24,19 @@ const ModifyCard = () => {
               <CCol lg="3">
                 <b>Role</b>
               </CCol>
-              <CCol>{account.is_admin? "Admin":"Member"}</CCol>
+              <CCol>{account.is_admin ? "Admin" : "Member"}</CCol>
             </CRow>
             <CRow>
               <CCol lg="3">
                 <b>Status</b>
               </CCol>
-              <CCol>{account.is_active? "Active":"Pending"}</CCol>
+              <CCol>{account.is_active ? "Active" : "Pending"}</CCol>
             </CRow>
             <CRow>
               <CCol lg="3">
                 <b>Registered</b>
               </CCol>
-              <CCol>{account.registered}</CCol>
+              <CCol>{registFormat(account)}</CCol>
             </CRow>
           </div>
         </CCol>
